[#161] di 문서 섹션 분리
